requirements.txt
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
cryptography==44.0.0
pyyaml==6.0.2
